Revert "fmt/pl: cast pl_i32/64 before negation"
This reverts commit c3ea09133d44b646338c9a2ff95d48999a6220cb.

+++ src/fmt/pl.c
@@ -91,7 +91,7 @@ int32_t pl_i32(const struct pl *pl)
 		}
 	}
 
-	return neg ? -(uint32_t)v : v;
+	return neg ? -v : v;
 }
 
 
@@ -132,7 +132,7 @@ int64_t pl_i64(const struct pl *pl)
 		}
 	}
 
-	return neg ? -(uint64_t)v : v;
+	return neg ? -v : v;
 }
